a shot with buckshot
